+++ flexiapi/resources/views/admin/account/index.blade.php
@@ -90,18 +90,15 @@
                     </td>
                     <td>
                         @if ($account->activated)
-                            <span class="badge badge-success" title="Activated">Act.</span>
+                            <span class="badge badge-success" title="{{ __('Activated') }}"><i class="ph">check</i></span>
                         @endif
                         @if ($account->superAdmin)
-                            <span class="badge badge-error" title="Admin">Super Adm.</span>
+                            <span class="badge badge-error" title="{{ __('Super Admin') }}">Super Adm.</span>
                         @elseif ($account->admin)
-                            <span class="badge badge-primary" title="Admin">Adm.</span>
-                        @endif
-                        @if ($account->sha256Password)
-                            <span class="badge badge-info">SHA256</span>
+                            <span class="badge badge-primary" title="{{ __('Admin') }}">Adm.</span>
                         @endif
                         @if ($account->blocked)
-                            <span class="badge badge-error">{{ __('Blocked') }}</span>
+                            <span class="badge badge-error" title="{{ __('Blocked') }}"><i class="ph">prohibit</i></span>
                         @endif
                     </td>
                     <td>{{ $account->updated_at }}</td>