+++ flexiapi/resources/views/admin/account/parts/tabs.blade.php
@@ -1,6 +1,7 @@
 @include('parts.tabs', [
     'items' => [
         route('admin.account.edit', $account->id) => __('Information'),
+        route('admin.account.external.show', $account->id) => __('External Account'),
         route('admin.account.statistics.show_call_logs', $account->id) => __('Calls logs'),
         route('admin.account.device.index', $account->id) => __('Devices'),
         route('admin.account.statistics.show', $account->id) => __('Statistics'),