+++ CHANGELOG.md
@@ -26,6 +26,7 @@ v1.7
 - Fix FLEXIAPI-261 Remove the TURN part in the XML provisioning (and only keep the API endpoint)
 - Fix FLEXIAPI-275 Add names in Spaces
 - Fix FLEXIAPI-278 Complete and reorganize the Markdown documentation
+- Fix FLEXIAPI-233 Add External Accounts (new version)
 
 v1.6
 ----